refactor/backend-structure
