+++ README.MD
@@ -1,0 +1,42 @@
+# 📡 LiveApp Backend
+
+This is the backend API service for **LiveApp**, a live streaming platform. Built with **Node.js**, **Express**, and **MongoDB**, it provides secure endpoints for user authentication, data access, and streaming control.
+
+---
+
+## 🚀 Features
+
+- JWT-based user authentication
+- MongoDB integration via Mongoose
+- Modular route/controller/service structure
+- Environment-based config with `.env`
+- API-ready structure for frontend integration
+
+---
+
+## 📁 Project Structure
+
+├── server.js
+├── .env.example
+├── src/
+│ ├── app.js
+│ ├── config/
+│ ├── controllers/
+│ ├── middleware/
+│ ├── models/
+│ ├── routes/
+│ └── services/
+
+
+---
+
+## 🛠️ Getting Started
+
+### 1. Clone the repository
+
+```bash
+git clone https://github.com/your-org/wobo.git
+cd wobo
+
+npm install
+npm start
