+++ server.js
@@ -1,15 +1,9 @@
-const express = require('express');
 const dotenv = require('dotenv');
-const connectDB = require('./config/db');
-const userRoutes = require('./routes/userRoutes');
+const connectDB = require('./src/config/db');
+const app = require('./src/app');
 
 dotenv.config();
 connectDB();
-
-const app = express();
-app.use(express.json()); // 处理 JSON 请求体
-
-app.use('/api/users', userRoutes);
 
 const PORT = process.env.PORT || 5000;
 app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
